requirements/base.txt
binaryornot==0.4.4
CacheControl==0.12.11
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.12
click==8.1.2
colorama==0.4.4
cookiecutter==1.7.3
idna==3.3
Jinja2==3.1.1
jinja2-time==0.2.0
lxml
MarkupSafe==2.1.1
msgpack==1.0.4
nanoid==2.0.0
poyo==0.5.0
Pygments==2.14.0
python-dateutil==2.8.2
python-slugify==6.1.1
pywebcopy==7.0.2
requests==2.27.1
six==1.16.0
sqlparse==0.4.3
urllib3
rollbar==0.16.3
text-unidecode==1.3
urllib3==1.26.9
zipp==3.12.0
django-permissions-policy==4.15.0
whitenoise==6.4.0
django-cors-headers==3.14.0
django-environ==0.10.0
django-mail-auth
django-htmx==1.14.0
django-extensions==3.2.1
django-crispy-forms==2.0
crispy-bootstrap5==0.7
django-simple-captcha==0.5.17
django-hijack==3.3.0
Django==4.2
arrow==1.3.0
django-throttle-requests==0.7.0
huey==2.5.0